noux: set initial time by using RTC session
Up to now Noux used a monotic clock whose initial start value always
was '0' (which correlates to the start of the UNIX epoch) to provide a
notion of time. In addition it is now possible to use the RTC session
to set the initial value for use-case where having a proper real world
time matters. It does not do any time-zone conversion - the returned
time depends on the component provide the RTC service.

To use the RTC session the 'rtc' attribute of the '<config>' node
must be set to 'yes'. Thereby the session becomes a mandatory
dependency as Noux will not start without it.