imx8q: experimental etnaviv driver port
The driver is ported from the vendor kernel 'linux-fslc-5.4-2.3.x-imx'
and tested on the iMX8MQ-EVK board. The SoC features a Vivante GC7000L.

The employed porting effort uses a hybrid approach that relies on the
newly developed 'dde_linux' tools. In particular 'create_dummies' was
used automatically generate dummy implementions for various Linux
functions and 'list_dependencies' to eventually gather all needed
contrib sources.

For providing the actual Linux emulation environment the old 'lx_kit'
was copied and adjusted were necessary. The 'lx_impl' layer that
implements the “user-facing” API, however, was omitted and replaced
by a local implementation of the missing functions. This is needed
because it will clash with using the vanilla Linux header files.
The seperation of the C code of Linux and the C++ code of Genode was
maintained by introduction an special-tailored C API ('lx_emul_*').