TEMPORARY ldso: remove mutex from dl_iterate_phdr
As a test remove the mutex from 'dl_iterate_phdr' as it is not
immediately clear if it is strictly necessary. Let us see how the
CI tests cope with that.

The issue was raised because the callback of the function retried
to relocate a jmp_slot that, which is also guarded by the same
mutex.