gpu_session: add buffer_handle() method
THe 'buffer_handle' method is used to get a local handle for the given
dataspace capability. It is used to provide matching handles for
translating between DRM and Gpu session in the case where the DRM API
is used in the front end (client) as well as the back end (driver).