gpu_session: introduce mapping type
The various mapping methods are modelled after the requirements of
the Intel GPUs or rather the Mesa driver back end.

With upcoming support for other driver back ends, we need to
sequeeze their requirements in as well. For now hijack 'map_buffer'
to provide for specifying the kind of mapping the client needs.

To make it clear that it is only needed in specific cases the default
mapping type is called 'IGNORED'.

Issue #4265.