pc: update wlan driver
The driver falls in line with the previous 'wifi_drv' component where
the ported wireless LAN stack and device driver is encapsulated in a
library. This library in return is used by the 'Libc::Component'
providing the necessary environment for the 'wpa_supplicant'.

In constrast to the old driver a 'wlan' VFS plugin is in charge of
initalizing the 'Lx_kit::Env' prior to executing any static
constructors.

Fixes #4455.