fixup 'dde_linux: initial wlan driver skeleton' (bring in frontend + wpa) -- DOES NOT COMPILE YET
This commit includes a copy of all of the user-space portions of the
current 'wifi_drv' component to cleanly seperate them for the time
being. The basic idea is to keep the current front end but retro fit
it to the new driver (not sure how yet, add a Fs session to the driver
and move the supplicant into its own self standing component comes to
mind…).