pc: add audio driver for HDA/SOF devices
This commit introduces a PC driver for HDA (Azalia) and SOF (DMIC)
audio devices. It is based on DDE Linux and its configuration is
similar to ALSAs mixer controls. As there is no auto-detection in
place yet the driver needs to be set up specifically for the
configuration of the system in question.

The driver supports common PCI HDA devices and systems making
use of SOF (DMIC) to enhance recording capabilities starting from
the Tiger Lake platform. Other companion devices, like DSP, are
currently not supported.

For more information on how to configured the driver please consult
the README file at 'repos/pc/src/driver/audio/pc/README'.

Fixes .