WIP wifi/pc: add RTL8188EE support
So far only [10ec:8179] (rev 01) was tested. Since the access to the
PCI config space is restricted (see XXX messages) the driver loads
the non power-saving FW and the driver port is therefor only tested
with that.

The 'wifi.run' run script works fine and Sculpt does too.
But you should not try to restart the driver and make sure your
AP is within reach (I think that expose another unhandled corner
case in the management layer *sigh*).