mixer: apply mute state while mixing
The 'mixer' did not mix channels that were muted in an attempt to
prevent unnecessary work. For wetter or worse this commit will
always mix all channels but apply the 'muted' state to calculate
the resulting volume, i.e., if the output channels are muted it
will produce a lot of zeros.