DRIVERS dde_linux: delay audio-jack sensing by 1000 ms
On Fuji-U7xxx the jack-plug events are received very early (jack half plugged).
When the plug is inserted too slowly, the headset-microphone recording might
start before the plug is in place, leading to permanent random noise. This patch
does not fix but reduces the problem by waiting 1s before starting the
recording.