base-hw imx epit-timer: fix duration calculation
The timer used to read the counter first and then the IRQ status. This could
cause a non-wrapped counter value to be considered a wrapped counter value,
leading to bogus timeout durations.

This commit fixes the bug and documents the used timer mode in the driver in
order to make future debugging of the driver easier.

Ref #4959