[WIP] os: add 'resolvconf' VFS plugin
This plugin consumes the content of the IP-stack's nameserver file
and produces 'nameserver x.y.z.w'. The file may be specified by
setting the 'nameserver_file' attribute, the default is
'/socket/nameserver'. The name of the provided file is specified
by setting the 'name' attribute, the default is 'resolv.conf'.