vfs/lxip: gather IP config via callback
The added callback mechanism allows for reacting upon changed IP
configuration information and is used to update the various
address files.

Since we now always know the actual size in advance reflect this
fact when providing status information. This allows for using
the VFS utilities to interact with these files.