sculpt_manager: move usb_hid claiming to config
The users may now interactively change the config for USB HID device claiming.
Beforehand, the users had to re-compile/re-build the Sculpt image, if they
wanted to assign a USB head set to a VM. Most often the Headset is also a
HID device and was already claimed by the usb_hid driver unfortunately.