lx_emul: set refcnt in __alloc_pages_bulk
Each page object ends up with an refcnt of 0, which apparently was
not a problem so far. However, it leads to memory issues when used
by the network's page-pool allocator as was triggered by its usage
in the FEC driver.

Due to uninitialized refcnt the allocations would not be recycled
and the page-pool allocator ended up allocating new caches that
led to out of resources requests.