vbox6: add system-state handler
With this commit VBox6 will react upon system-state changes. If the
'consider_system_state' config attribute is set to 'true' (default is
'false) VBox6 opens a connection to a 'system' ROM and depending on
the 'state' attribute will either attempt to reset or power-down the
VM.

This allows for managing the life-cycle of a VM without relying on a
GUI session.