DRIVERS sculpt_manager: add suppressing of storage devices
This commit adds support for instructing the 'sculpt_manager' to
refrain from starting the drivers for storages devices like NVMe and
AHCI controllers. It is then up to the preset or rather deploy config
in question to manage such devices directly.

The following snippet shows how The 'manager' configuration has to be
extended to supress NVMe as well as AHCI from being managed by the
'sculpt_manager':

!<config ahci="false" nvme="false"/>