lighttpd: prevent invalid fdarray access
The implementation relies on the used fds to be densely packed, which
is not the case with our libc. Since the returned fds were much larger
than expected because the libc itself consumes multiple fds when
managing sockets using the fd to directly addressing an array elment
led to memory corruption.