DRIVERS nitpicker: deliver press events to global key receiver
This is a quickfix for the following misbehavior: If a global key is
pressed (e.g., Caps Lock in Sculpt) and the mouse is clicked before
release, the global-keys handler receives the global-key press but no
further events of the sequence.

A more reasonable fix shall address the "else branch" that deletes the
_input_receiver in case of mouse events without a hovered client too,
which is beyond the narrow scope of this quickfix.