WIP sculpt_manager: omit keyboard focus when rfkilled (see note at the end of the commit msg)
The keyboard focus selection mechanism currently does not take into
account when the 'wifi' driver is in the 'rfkilled' state and in this
situation prevents the user from interacting with the inspect window.

As Sculpt does not allow for setting the 'rfkill' option via the GUI
this issue only becomes apparent when the 'wifi' configuration is
manually managed by the user.

NOTE: at the moment it is not possible to start the driver with
      rfkill enabled because it will be applied at a time where
      the driver is not yet ready (the module is not loaded and
      switching TYPE_WLAN off gets lost). So this commit could
      lead to starting users with the driver “rfkilled” when it
      is actually still on.