vfs/block: set default I/O buffer to 4 MiB
This commit raises the Packet_stream bulk-size to 4 MiB from the
128 KiB used before. It also removes the 'block_buffer_count' option
as the count is now calculated internally and covers the complete
bulk-size. A new 'io_buffer' option is introduced that manages the
bulk-size.

This change can double the throughput with sequential I/O operations
when employing larger block-sizes.

Issue #2263.