fixup! hw: simplify x86 vCPU initialization
Set the STARTUP_EXIT from the Hw_vcpu constructor and move its
definition out of the kernel. Check the vCPU's initialization state on
x86_64 at the level of the `pause()` system call and remove the
`_sync_from_vmm()` and `_sync_to_vmm()` hooks that are superseeded by
specializing `pause()` for x86_64.

Fixes #5483