app/extract: limit supported formats
This commit limits the support to the formats that are in active
use at the moment:

  * tar and raw are required for Sculpt (archives and index)
  * zip is used by Genodians

Instructing libarchive to use all supported formats can lead to
unexpected results when the probing mechanism identifies the
wrong format. For example, an empty index file ('<index> </index>')
will be classified as MTREE format.

Fixes #5528.