os/packet_stream.h: propagate construction errors
The initialization of the packet-stream allocator via
'Range_allocator::add_range' can fail with an 'Alloc_error'. This could
be observed with run/nic_router_stress running in Pistachio.

The patch reflects the construction error as constant 'constructed' that
can now be evaluated by the creating code.

It also fixes the particular problem of the NIC/Uplink-session creation
error in the NIC router and adds diagnostic messages to reveal further
manifestations of this problem.

Fixes #5566