sel4: avoid using same id for gsi/msi allocation
Up to now it may happen, that the chosen vector used for MSI is the same
for two drivers, when the both drivers used the same IRQ as parameter
during IRQ session construction. The IRQ value is not relevant for MSI
construction actually, tough allocate a MSI vector dynamically from a
valid range.

Issue #5581