nvme: implement SQ multiplexing
This commit introduces SQ multiplexing in the driver. By setting the
'force_sq' attribute of the 'config' node to 'yes' the component is
instructed to create one I/O queue only and to handle concurrent
clients in a alternating fashion. This mode of operation is in line
with the other block device driver components that do not support
multiple I/O queues.

A more thorough implementation should handle multiplexing multiple
clients on one I/O dynamically. As of now we lack the practical
experience when using multiple I/O queues (for concurrent clients,
which may or may be not the intended use-case).

Issue #5479.