hw: set IA32_TSC_AUX MSR for vCPUs and threads
The IA32_TSC_AUX is returned by `rdtscp` and typically used to detect
the CPU id that the TSC was taken on. For vCPUs, the TSC_AUX value can
be explicitly set by the VMM. We therefore need to change the MSRs value
to the guest's version and vice versa when context-switching.

Issue #5639