menu_view: define label text as sub-node content
This patch changes the way of specifying the text of label widgets from
the use of a 'text=' attribute value to the use of quoted content hosted
in a <text> sub node. This way, leading/trailing whitespace can be
expressed as quoted content without relying on the support of
leading/trailing spaces in attribute values.

Issue #5636