core: give mapped_ram to platform services & pd
This commit hands over the 'Accounted_mapped_ram_allocator' of a
pd to the constructor of 'Platform_pd' to enable kernel-specific
code to make use of it, e.g. to allocate page-tables with it.

Moreover, the 'Mapped_ram_allocator' of core is handed over to
the initialization of platform-specific services, e.g. Vm service.

Issue #5645