base: orderly roll back Thread::stack_size effects
Upon the successive expansion a thread's stack by calling
'Thread::stack_size', the RAM allocations for the expansions went
unrecorded and could thereby not be released on destruction. This patch
records up to four RAM mappings per stack and deallocates the RAM used
for stack expansions.

We haven't observed this limitation as a problem in practice because
'Thread::stack_size' is only used internally be the dynamic linker for a
thread that is never destructed. But scenarios that spawn and kill many
threads using large stacks would leak RAM without this patch.

Fixes #5683