sculpt_manager: handle WPA3-only networks
This commit allows for connecting to WPA3-only protected wireless
networks. The previous implementation obnly allowed for connecting
to WPA2/WPA3 mixed networks and chose WPA2 as protection scheme.

Since the key management differs between WPA1/WPA2 and WPA3 and thus
the interaction with the supplicant is different as well this must
be handled explicitly.

Fixes #5688.