sculpt_manager: update focus after key release
When entering a wifi password and pressing KEY_ENTER, the focus
report will be updated immediately. In consequence, KEY_ENTER may
unintentionally be submitted to the inspect view or file view as well.
By only updating the focus report when no key is pressed, we circumvent
the race between input events and focus report.

Fixes #5532