vfs: reflect eof when reading a dir as a file
With this patch, reading from a dir handle no longer returns
default-constructed dir entries but a read-count of zero, signalling an
end-of-file to the caller. This way, 'cat' no longer loops infinitely
when passing a directory as argument.

Besides reflecting EOF in the dir file system, the os/vfs.h utility for
reading directory entries had to be adjusted to accept a read count of
zero without printing a warning.

Fixes #5514