driver/audio: always sent progress signal
This commit changes the way in which the PC audio drivers operate
in that these now always submit a progress signal, even if they
are only playing silence.

This in turn keeps the 'mixer' component running that then will
check each session as in the meantime clients might have queued
additional packets.