wm: suppress pointer updates when typing
The pointer report is generated when all keys are released because this
may be the end of a drag operation. However, if the pointer position
did not change, there is no point in updating the report. Suppressing
these non-updates reduces a lot of noise at the decorator when typing.

Issue #5716